Add DeleteTodo endpoint to TodosController
Introduced a new HTTP DELETE endpoint, `DeleteTodo`, in the
`TodosController` class. This method accepts an `id` parameter
to locate a `todo` item in the `_todos` collection. If the item
is not found, it returns a `404 Not Found` response with an
appropriate message. If the item is found, it removes the item
from the collection and returns a `204 No Content` response.